requirements.txt
matplotlib==3.5.3
pandas==1.4.4
Pillow==9.3.0
streamlit==1.13.0
tensorflow==2.10.0
